db_populator/requirements.txt
# Python dependencies
beautifulsoup4==4.12.2
pymongo==4.6.1
selenium==4.16.0
python-dotenv==1.0.0
pyyaml==6.0.1
